requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
pytz==2023.3
qrcode==7.4.2
Pillow==10.1.0
requests==2.31.0
Flask-Compress==1.14
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
redis==5.0.1
gunicorn==21.2.0
psutil==5.9.6
APScheduler==3.10.4
celery==5.3.4